fix des erreurs de stop

+++ watchdog.py
@@ -10,6 +10,13 @@ from worker import serveur_worker
 dw_pipe_path = 'pipes/dwtube'
 wd_pipe_path = 'pipes/wdtube'
 
+already_killed = False
+
+if os.path.exists(dw_pipe_path):
+    os.remove(dw_pipe_path)
+if os.path.exists(wd_pipe_path):
+    os.remove(wd_pipe_path)
+
 # Crée les dossiers et pipes si besoin
 if not os.path.exists('pipes'):
     os.makedirs('pipes')
@@ -18,14 +25,23 @@ if not os.path.exists(dw_pipe_path):
 if not os.path.exists(wd_pipe_path):
     os.mkfifo(wd_pipe_path)
 
+if not os.path.exists('hearthbeats'):
+    os.makedirs('hearthbeats')
+
 child_pids = []
 
 def kill_children(signum, frame):
+    global already_killed
+    if already_killed:
+        return
+    already_killed = True
     print("Watchdog: Terminaison en cours...")
     for pid in child_pids:
+        os.remove('hearthbeats/server_' + str(pid))
         os.kill(pid, signal.SIGTERM)
     os.remove(dw_pipe_path)
     os.remove(wd_pipe_path)
+    
     sys.exit(0)
 
 def check_alive(pid):
@@ -41,9 +57,12 @@ def launch_server(func, args_tuple, child_pids):
         print("Erreur lors du fork.")
         sys.exit(1)
     elif pid == 0:
+        signal.signal(signal.SIGTERM, signal.SIG_DFL)
+        signal.signal(signal.SIGINT, signal.SIG_DFL)
         func(*args_tuple)
     else:
         print(f"Watchdog: J'ai lancé un serveur avec le PID {pid}.")
+        open('hearthbeats/server_' + str(pid), 'w').close()
         child_pids.append(pid)
         return pid
 
@@ -59,10 +78,11 @@ if __name__ == "__main__":
     pipe_in_wdtube = os.open(wd_pipe_path, os.O_RDONLY | os.O_NONBLOCK)
     pipe_out_wdtube = os.open(wd_pipe_path, os.O_WRONLY)
 
+
     dispatcher_pid = launch_server(serveur_dispatcher, (shared_mem, pipe_out_dwtube, pipe_in_wdtube), child_pids)
     worker_pid = launch_server(serveur_worker, (shared_mem, pipe_in_dwtube, pipe_out_wdtube), child_pids)
 
-    print("Watchdog: J'ai lancé les serveurs.")
+    print("Watchdog: Serveurs lancés.")
 
     while True:
         if not check_alive(dispatcher_pid):
